edit: api working, repl display
