+++ src/ui/repl.js
@@ -1,9 +1,9 @@
 import blessed from 'blessed'
-import { getPrices, getQuote } from '../api/api.js'
 
-export function buildRepl(row, col, h, w) {
-  const self = this
-  const history = `{#2ea-fg}Welcome to iexcli.{/}
+import { getPrices, getQuote } from '../api/api.js'
+import { help } from './help.js'
+
+const history = `{#2ea-fg}Welcome to iexcli.{/}
   Data provided by IEX Cloud 
   {blue-fg}<https://iexcloud.io>{/}
 
@@ -13,12 +13,14 @@ export function buildRepl(row, col, h, w) {
   {underline}{#4be-fg}<https://github.com/hp4k1h5/iexcli>{/}...
 \n\n\n `.split('\n')
 
-  const repl = this.grid.set(row, col, h, w, blessed.form, { keys: true })
+export function buildRepl(ws, c) {
+  const repl = ws.grid.set(...c.yxhw, blessed.form, { keys: true })
 
   // console display (optional), otherwise commands just have effects and don't report
   const output = blessed.textarea({
     parent: repl,
-    height: '80%',
+    name: 'output',
+    height: '75%',
     tags: true,
     style: {
       scrollbar: {
@@ -29,11 +31,11 @@ export function buildRepl(row, col, h, w) {
   // init welcome text
   output.setValue(history.join('\n'))
 
-  // non-optional all interaction is handled here
+  // all interaction is handled here
   const input = blessed.textbox({
     parent: repl,
     name: 'input',
-    top: '80%',
+    bottom: 0,
     height: 3,
     inputOnFocus: true,
     border: { type: 'line' },
@@ -51,7 +53,7 @@ export function buildRepl(row, col, h, w) {
   })
   repl.on('submit', async function (data) {
     // parse and handle input
-    let evaluation = await evaluate(data.input, self)
+    let evaluation = await evaluate(data.input, ws)
 
     // mimic scroll
     history.push('{bold}> {/}' + data.input)
@@ -61,13 +63,13 @@ export function buildRepl(row, col, h, w) {
     // clear input and refocus
     input.clearValue()
     input.focus()
-    self.screen.render()
+    ws.screen.render()
   })
 
   input.focus()
   // add to curScreen
-  this.repl = repl
-  this.screen.render()
+  ws.repl = repl
+  ws.screen.render()
 }
 
 // helpers
@@ -105,57 +107,6 @@ function exit(self) {
     console.log('exiting...')
   }, 1000)
   return '{#abf-fg}goodbye...{/}'
-}
-
-function help(self, words) {
-  // handle help arguments if any
-  let what = words[words.findIndex((w) => /h(elp)?/.test(w)) + 1]
-  if (what) {
-    const whatSym = `
-    {bold}{#2ea-fg}help {#cd2-fg}\${/} (symbol)
-prefix stock ticker symbols with {#cd2-fg}\${/} to change the active symbol. Including a $-prefix will automatically update all charts and tables
-ex. {#cd2-fg}$brk.b{/}
-ex. {#cd2-fg}$GOOG{/}`
-    const whatTime = `
-    {bold}{#2ea-fg}help {#cd2-fg}:{/} (time range)
-prefix valid time ranges with {#cd2-fg}:{/} to change the active time range. Including a :-prefix will automatically update all the charts and tables
-{bold}{#2ea-fg}valid time ranges:{/}
-${self.validUnits.map((u) => '{#cd2-fg}:' + u + '{/}').join(' ')}
-AND numbers with minute {#cd2-fg}min{/} or hour {#cd2-fg}h{/}
-ex. {#cd2-fg} :100min {/}{#ddd-fg} last 100 minutes of aggregated trading data{/}
-ex. {#cd2-fg} :6.5h {/}{#ddd-fg}{/}  last 6.5 hours of minute aggregated trading data`
-    const whatExit = `
-    {bold}{#2ea-fg}help {#cd2-fg}exit | quit{/} (time range)
-exits iexcli`
-
-    const whats = {
-      $: whatSym,
-      ':': whatTime,
-      exit: whatExit,
-      quit: whatExit,
-    }
-    return whats[what] ? whats[what] : `{red-fg}error{/}: no help for ${what}`
-  }
-
-  // return full help
-  return `
-    {bold}{#2ea-fg}help menu{/} 
-{bold}available commands:{/}
-{#cd2-fg}h(elp){/}   :prints this menu
-{#cd2-fg}\${/}        :ticker symbol prefix
-          changes active symbol
-          ex. {#cd2-fg}$qqq{/}
-          try {#cd2-fg}help \${/}
-{#cd2-fg}\:{/}        :time (range) prefix
-          changes active time range
-          ex. {#cd2-fg}:6m{/}
-          try {#cd2-fg}help \:{/}
-
-these commands can be aggregated:
-    ex. {#cd2-fg}$z :10h{/}
-    ex. {#cd2-fg}:1y $GM{/}
-
-{#cd2-fg}exit / quit{/}     :quit iexcli`
 }
 
 async function update(self) {
